// src/types.ts
export interface EnquiryFormData {
  name: string;
  place: string;
  cateringType: string;
  approxPerson: string;
  description: string;
  phone: string;
  email: string;
}

export interface Occasion {
  title: string;
  image: string;
  description: string;
}

export interface Service {
  title: string;
  icon: string;
  description: string;
}